Rename sqlite_test to test_sqlite
This avoids naming conflicts when we want to write unit tests for a sqlite package.
